chore: reorg - move backend files to root folder
