+++ package.json
@@ -2,7 +2,6 @@
   "name": "backend",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
   "type": "module",
   "directories": {
     "test": "test"
